SERVER-4312 rewrite keyboard parsing for Linux
Replace the nested if statement with a ton of new code to parse
the escape sequences that the various Linux terminal programs
give us when arrow keys (for example) are pressed while Ctrl or
Alt are held down.  Make ESC followed by <key> generate the same
internal code as Alt+<key> even when the terminal gives us different
stuff.  Meta-B and Meta-F now work on all platforms.  Rip out the
unsupported terminal "feature", it does nothing useful.  Turn on
the Linux keyboard debugging feature in all builds, it is useful
and doesn't get in the way.  Striving to work correctly on Gnome
Terminal, xterm, aterm, konsole, rxvt and yakuake, code was written
with this set targeted but testing is limited to implemented
functions (e.g. Meta-B and Meta-F).  More testing once more features
are written.