avoid deadlock on shell exit
The mongo shell has the ability to run other programs to facilitate unit testing.
It spawns a thread to shovel output from the spawned program to the terminal via a pipe.
At process exit time, the thread may continue to run and attempt to use std::ostreams,
which on Windows may deadlock if the internal critical section is destructed while you're
using them.
The output threads do check a dbexit flag and use that to determine when to stop outputting.
This commit fixes a race condition between the setting of that flag and the thread exit.