SERVER-4892 - Perform correct authorization checks when doing killCursors.
Prior to this patch, we were incorrectly interpreting a portion of the dbKillCursors message as a
namespace string, for auth purposes.  This was nonsense, as each cursor listed in the killCursors
could be from a distinct namespace.  In this patch, the client's authority to kill each cursor is
checked as part of ClientCursor::erase.  Failure to erase due to lack of authorization is reported
identically to failure because the cursor never existed or no longer exists.

Test to follow.