SERVER-5152 Improve unhandled exception code on Windows
This change improves the reporting of unhandled exceptions in
mongod.exe in Windows, mainly access violations (segfaults).  Like
the old code, it doesn't try to keep running, but in release
builds this version does try to use dbexit() to exit instead of
letting Windows do it to us.  In debug builds, we pass the exception
on to Windows as before, which lets it offer the chance to run the
debugger.  Tested with a null pointer reference, we now get a clean
log with a thread name and the address where the exception happened.