SERVER-5685 multiple threads prefetch and apply ops on secondaries
On a secondary, ops are now pulled off the network in one thread and
placed in a deque. Another thread pulls off the ops in batches. These
batches are presented to a prefetch thread pool (which prefetches all
the pages for index traversal and data), and a writer thread pool
(where operations are applied to the database).  Each op is assigned a
thread based on a hash of its namespace, in order to maintain write
order within a given document.  During the write phase, all readers
are blocked (even if a given write hits a yieldpoint).
Write concern is updated on the primary by using a separate query
cursor.