SERVER-5211 Separate of CmdLine::store into two functions.
The new functions are CmdLine::store, with the same signature as before, and
mongo::initializeServerGlobalState(), which should take no parameters, but does
take an optional bool to deal with a knot in mongod --shutdown that I haven't
yet untangled.

This separation makes it easier to test functions in the command line parsing
part of the library, and provides a little extra clarity on the role of the two
functions.