SERVER-7700 Use high-resolution program uptime to track socket connection time
Switch from curTimeMicros64() to new getProgramUptimeMicros() function for
tracking socket connection time.  curTimeMicros64() on Windows uses
GetSystemTimeAsFileTime(), which has a resolution of 1 to 15 milliseconds,
which is not good enough for tracking fast events.  getProgramUptimeMicros()
provides program uptime instead of clock time, but uses QueryPerformanceCounter()
on Windows and so has sub-microsecond resolution.