SERVER-7202 SSL certificate validation in both directions
New command line flag --sslCAFile specifies a file containing certificates
for the certificate authority. This flag is supported on mongod, mongos, and mongo
shell.

mongo shell now supports --sslPEMKeyFile and --sslPEMKeyPassword to specify the
client certificate presented as part of the SSL handshake when connecting to a server.

Note that certificate validation is not forced; validation only currently occurs
if the client presents a certificate.  Note that CRL or OSCP remains unimplemented,
and there is not yet a way to restrict the cipher used.

Note that all of these features require --ssl parameter passed to scons at build time.