SERVER-11995 Have authSchemaUpgrade sent to a mongos upgrade auth schema everywhere.
In particular, sends the commands to each shard, as well as performing the upgrade on the
config servers.

Also, allow user to specify maximum number of upgrade steps to take on any given shard.
Defaults to "as many as necessary".

To achieve these changes, separate implementations of the upgrade command are provided to
mongos and mognod.