Do not zero-fill newly allocated files on Windows
On Windows platforms code was added to zero-fill each newly allocated data file in order to work around Windows KB http://support.microsoft.com/kb/2731284. The KB only applies for Windows 7 and Windows 2008 Server installations, so there is no need to do it on each and every system.

This fix restricts the zeroing-out code to only versions, which have had the bug. It will not check whether the hot fix has been installed.