SERVER-13631 Bring mongodump --repair implementation under RecordStore
This change adds the concept of "repair" iterator, which is used in the
event of data corruption. It has the ability to walk corrupted record
store and salavage as much data as possible.

Most of the functionality under mongodump --repair has been moved in there
with mongodump remaining as just a driver.