SERVER-16065 MMAP V1 remap always
We cannot remap selectively, because we now do not have a way to apply
back-pressure on the inserts coming in while journal I/O is happening
outside of the flush lock. That way if the journal I/O (WRITETOJOURNAL)
takes long time, we end up accumulating too much data in memory and
triggering rate limiting wassert.