SERVER-19111 Add ability to predeclare mock targeters
In certain cases, namely ShardRegistry::createConnection, we request a
connection to be created to a host which is not yet a shard, but should be
talked to via the shard connection object.

In these cases we have no control over the mock targeter which gets
created.

With this change we allow targeters to be pre-declared.