SERVER-19001 Implementation of ThreadPoolTaskExecutor and basic tests.
In order to support deterministic unit testing, it was also necessary to
introduce a mock implementation of ThreadPool, executor::ThreadPoolMock,
that is tightly integrated with NetworkInterfaceMock to allow for
deterministic unit testing of things that use TaskExecutors.

To keep the ThreadPoolTaskExecutor from having to keep a dedicated
thread for handling scheduleAt(Date_t, ...) processing, a new method,
setAlarm, is introduced to NetworkInterface. setAlarm offers an
extremely relaxed contract, to maximize the number of legal implementations.