SERVER-20019 SpiderMonkey performance improvement in shell by removing pre-emptive string creation
The current SpiderMonkey implementation preemptively generates a string
for the IdWrapper of JSid. This id->std::string conversion is expensive
and is quite often unnecessary. By delaying the creation of the string
until after checking that it is absolutely necessary, we can improve
SpiderMonkey performance on the shell by about 20%.

Signed-off-by: Adam Midvidy <amidvidy@gmail.com>