SERVER-27243 Ignore invalid collection options if 'create' present
It is possible for older versions of MongoDB to save invalid collection
options which should be ignored. Older versions also would include a
'create' field in the collection options when a collection was created,
and possibly re-order the fields of the options document during a
subsequent 'collMod'. If the 'create' option is present at all, we
should take that to mean that these options were created on an older
version of MongoDB, and ignore unknown options. Otherwise, we should
reject unknown options.

(cherry picked from commit 6fba074768fad5f47611de38682257257005a1a6)