SERVER-27800 Change ReplicationCoordinatorImpl and ReplicationExecutor constructors to take owned objects by unique_ptr
The constructors of ReplicationCoordinatorImpl and ReplicationExecutor
predate C++11 support in the MongoDB codebase, and so receive bare
pointers to objects that they actually own. This patch is to changes the
constructors and some of the associated test infrastructure to
transfer ownership by passing unique_ptr, as in more modern
components.