SERVER-28239 fix double close in ticket holder
The legacy transport layer allowed end() to be called multiple times on
the same session, which would incorrectly double close + double
increment the ticket holder.  This occured during primary step down, as
endAllSessions was called (which called end()) and then the owning
threads called end() as they finished.

Fix involves a mutex per connection and a check for _closed in
closeConnection.