SERVER-29277 Automatically retry operations on DatabaseDropPending.
Overrides the Mongo.prototype.find(), Mongo.prototype.runCommand(), and
Mongo.prototype.runCommandWithMetadata() functions to automatically
retry (for up to 10 minutes) on a DatabaseDropPending error response.
The override is loaded in for only the FSM worker threads when running
with replication and therefore doesn't impact the runner's connection
or non-replica set configurations.