SERVER-28765 Create BitNode.
Additionally, this change fixes a couple of minor things. First, it
makes the single-argument constructor for ArithmeticNode
explicit. Second, it adds a (noop = false) assignment in the else
branch of a test in set_node.cpp, instead of relying on the fact that
callers set noop to false. That gives SetNode clearer behavior.