jseyster/json-schema-feature-compatibility-2
If a user updates a 3.4 node to 3.6 but keeps the feature
compatibility version at 3.4, it should remain possible to be in a
replica set with 3.4 nodes or even to downgrade back to 3.4. To that
end, we ensure that is not possible to create a collection validator
with a $jsonSchema expression unless the feature compatibility version
is 3.6. A $jsonSchema validator would not replicate correctly to a 3.4
node, and its existence in the database would prevent a downgrade to
3.4.