SERVER-31827 Wait for Secondary after stepdown.
The primary being told to step down will kill all incoming connections
during that stepdown process.  When this happens, connections which
are incoming may be closed before they can enter a state which
"reserves" them from closure.  To facilitate correct semantics in this
test, we should wait for the primary to completely step down to a
secondary, before trying to authenticate to it again.