SERVER-31730 Preserve getNext status in PlanExecutor::executePlan().
It is possible for an update that gets interrupted because a primary
node stepped down to result in an error return from
PlanExecutor::getNext(), which ultimately propagates to the
client. Clients should know to retry any operation that fails for this
reason, but in PlanExecutor::executePlan(), the status code gets
rewritten to be OperationFailed, which clients do not know to retry.

This patch preserves the original error code, so that clients do the
right thing.

commit f6581038d95314794fcfc84f24253653a1ff861e
Author: Justin Seyster <justin.seyster@mongodb.com>
Date:   Fri Oct 27 16:56:59 2017 -0400

    SERVER-31730 Preserve getNext status in PlanExecutor::executePlan().