SERVER-33330 closeAll now checks for bg jobs in progress
Previously, closeAll() would have a "force" mode that caused it to either ignore bg jobs
in progress (force on), or avoid closing databases with bg jobs in progress.  This parameter
is no longer useful, so I have removed it.  The test-only restartCatalog command will now
return an error instead of crashing if you call it with bg jobs in progress.