SERVER-34960 Add MONGO_FAIL_POINT_BLOCK_IF
It's sometimes useful to be able to check a pre-condition on a fail
point without manipulating the fail point state (decrementing nTimes for
instance).

Adding a callable to shouldFailOpenBlock and shouldFail, and threading
that through a new block macro MONGO_FAIL_POINT_BLOCK_IF, allows for
observation of the fail point payload and a chance to abort without
extra overhead if the fail point is disabled.