SERVER-34682 Old primary should be able to store last vote when casting vote in new term
This patch allows a current primary to step down and cast its vote for a
new primary in a higher term in one step. This allows for a new
candidate to become elected faster if it relies on the vote of an old
primary. Previously, when learning of a higher term via a
replSetRequestVote request, a primary would step down, causing us to
interrupt the processing of the vote request when we try to acquire a
lock to write down our LastVote document with an
InterruptedDueToReplStateChange code. This patch allows us to ignore
these interrupts, so we can proceed with processing the vote request.

Additionally, this patch also asserts that the vote response is actually
ok, so that we don't erroneously send a voteGranted=yes response even if
we failed to persist our last vote document.