SERVER-35634 view cycle test should not always assert that finds will succeed
The aggregate command resolves all views to their backing collections,
then re-runs the command so that no views are present. However, it
releases the collection lock before creating and iterating the cursor.
In this window of time, one of the involved collections may be dropped
and replaced with a view.

This fix robustifies view_cycle_catalog_lookup.js so that it tolerates
this situation, as it is expected for the command to fail.

(cherry picked from commit 3406454039941a8ec1ff700fb6c3377ec6e5c5bc)